Add transition color on hover index

+++ removeproject.php
@@ -2,7 +2,10 @@
 require('config.php');
 ?>
 <?php
-
+if (!empty($_SESSION['id'])) {
 $removeproject = $bdd->exec("DELETE FROM projects WHERE id =" . $_GET['project'] . "");
 header('location: index.php');
+} else {
+  header('location: index.php');
+}
 ?>
